requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-groq==0.0.1
chromadb==0.4.22
streamlit==1.29.0
pandas==2.1.4
python-dotenv==1.0.0
sentence-transformers==2.2.2
huggingface-hub==0.20.1